app/requirements.txt
streamlit
gunicorn
pandas
matplotlib
seaborn
plotly
numpy
scikit-learn
python-dotenv
gcsfs
joblib
openpyxl

# itertools

# cloud gcp
google-cloud-bigquery 
db-dtypes
google-cloud-aiplatform
